+++ ChangeLog.md
@@ -8,6 +8,7 @@
   repository for viewing.
 - Added the ability to delete a single item of history from the history
   list.
+- Added the ability to clear down the whole of history.
 
 ## [0.4.0] - 2023-05-03